% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{Avg_IDW}
\alias{Avg_IDW}
\title{Inverse distance weighting}
\usage{
Avg_IDW(vals, weights)
}
\arguments{
\item{vals}{numeric vector. Values to compute average of.}

\item{weights}{numeric vector. Weights (proximities) for each value.}
}
\value{
numeric representing the weighted average of vals by weights
}
\description{
Small utility function to
}
