% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_utilities.R
\name{computeDist}
\alias{computeDist}
\title{Compute pairwise distance between single cells in PCA space}
\usage{
computeDist(sce, usePCs = TRUE, numPCs = 10)
}
\arguments{
\item{sce}{viccc object. Must have already executed the function runPCA().}

\item{usePCs}{logical. Whether to use PCs instead of gene expression values for distance computation.
Setting to FALSE may drastically increase runtime in datasets with many genes. Default TRUE.}

\item{numPCs}{numeric. How many PCs to use for distance computation. Larger values will take
longer, especially for large datasets, but also incorporate more information. Default 10.}
}
\value{
viccc object with defined grid points in PCA space
}
\description{
Computes grid points for future steps that will smooth the inferred velocities
in PCA space.
}
