% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{computeVector}
\alias{computeVector}
\title{Compute single-cell transition vector}
\usage{
computeVector(
  sce,
  query_point,
  useGinv = F,
  v2 = T,
  invertV2 = F,
  maxNeighbors = 100
)
}
\arguments{
\item{sce}{sticcc object}

\item{query_point}{character. Rowname of posMat indicating which sample to compute a vector for.}

\item{useGinv}{logical. Whether to use generalized inverse for computing least-squares regression. 
Default FALSE, but may be helpful with very homogeneous datasets.}

\item{v2}{logical. If TRUE, will calculate v2 (incoming transition) as well as v1 (outgoing). Default TRUE.}

\item{invertV2}{logical. If TRUE, will multiply v2 by -1 such that it points "forward in time" as v1. Default FALSE. Note this step
may be necessary for downstream analysis.}

\item{maxNeighbors}{numeric. Maximum number of neighbors to include in linear regression - if there are more neighbors in range 
than this limit, this value will be used to randomly sample a subset.}
}
\value{
named list of results containing X and Y (position on first two dimensions), boolean values for logging purposes, 
and predicted vectors.
}
\description{
Calculates predicted transition vectors v1 and v2 for a query point.
}
