% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_main.R
\name{DCComputeTrajectorySCE}
\alias{DCComputeTrajectorySCE}
\title{Infer velocities of single cells}
\usage{
DCComputeTrajectorySCE(sce)
}
\arguments{
\item{sce}{viccc object. Must have already supplied a topology and executed the functions runPCA() and computeDist().}
}
\value{
sce object with velocities - see colData(sce) for the first two components (dX and dY),
and sce@metadata$vectors for all components.
}
\description{
Computes the inferred velocity of each cell in the dataset according to the topology supplied previously.
}
