% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{distfun}
\alias{distfun}
\title{Distance function}
\usage{
distfun(x)
}
\arguments{
\item{x}{numeric input matrix.}
}
\value{
dist
}
\description{
Small utility function to compute dist of a given matrix
}
