% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{getNeighbors}
\alias{getNeighbors}
\title{Get nearby samples for a given query sample.}
\usage{
getNeighbors(sce, queryPoint, neighborhoodRadius)
}
\arguments{
\item{sce}{sticcc object}

\item{queryPoint}{character or vector. Either a rowname of sce, or a vector to be compared to 
the normalized expression data.}

\item{neighborhoodRadius}{numeric. Proportion of the maximum pairwise distance within sce to find
neighbors.}
}
\value{
A list containing the following elements: "neighbor.ids": the names of the nearby samples,
"neighbor.dists": distances from each neighbor to the query point.
}
\description{
Finds neighboring samples around the supplied point.
}
