% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{listToNumeric}
\alias{listToNumeric}
\title{Convert list to numeric}
\usage{
listToNumeric(x)
}
\arguments{
\item{x}{input list}
}
\value{
list converted to numeric data type
}
\description{
Small utility function to ensure a list of data is in numeric format
}
