% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_utilities.R
\name{loadTopo}
\alias{loadTopo}
\title{Load topology file}
\usage{
loadTopo(topoName, topoDir = NA)
}
\arguments{
\item{topoName}{character. Filename of a .tpo topology table (do not include file extension).}

\item{topoDir}{character. Path to the directory containing the .tpo file. Defaults to "input_topos"
and will create this directory if needed.}
}
\value{
data.frame of a network topology.
}
\description{
Returns a data.frame from reading a .tpo file with 3 columns: Source, Target, Type. Type 1 indicates
activation and type 2 indicates inhibition.
}
