% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot network to file}
\usage{
plotNetwork(sce, outputDir = NA, plot_suffix = NA)
}
\arguments{
\item{sce}{sticcc object}

\item{outputDir}{character. Directory to save plot. Default is a subdirectory of the working 
directory with the same name as sce@metadata$topoName, which will be created if it does not exist}

\item{plotSuffix}{character. String to append to plot filename. Default is the current time.}
}
\description{
Saves an image of the network topology
}
