% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_utilities.R
\name{prepMetadata}
\alias{prepMetadata}
\title{Create metadata table.}
\usage{
prepMetadata(exprMat, cluster = T, ...)
}
\arguments{
\item{exprMat}{data.frame or matrix. Non-normalized expression matrix with rows as features
and columns as cells.}

\item{cluster}{logical. Whether to perform hierarchical clustering.}

\item{...}{Additional arguments to the (internal) clustering function - k (number of clusters),
features (which features to cluster by, defaults to all), and clustfun (default "ward.D2")}
}
\value{
data.frame with cells as rows and metadata characteristics as columns.
}
\description{
Returns a data.frame containing metadata for each cell in a dataset. This can be
substituted with a previously existing data.frame and is only necessary if no metadata table
exists. This will also divide the dataset into k clusters using the ward.d2 method.
}
