% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_utilities.R
\name{runPCA}
\alias{runPCA}
\title{Perform PCA on gene expression data}
\usage{
runPCA(exprMat)
}
\arguments{
\item{exprMat}{data.frame or matrix. Expression matrix with rows as features and columns as
cells.}
}
\value{
list with class "prcomp" containing elements sdev, rotation, x, center, and scale.
}
\description{
Returns a vector of cluster numbers corresponding to each cell in a dataset.
}
