% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_main.R
\name{runSTICCC}
\alias{runSTICCC}
\title{Infer velocities of single cells}
\usage{
runSTICCC(sce, v2 = TRUE, useGinv = FALSE, ...)
}
\arguments{
\item{sce}{sticcc object. Must have already supplied a topology and executed the functions runPCA() and computeDist().}

\item{v2}{logical. If TRUE, will calculate v2 (incoming transition) as well as v1 (outgoing). Default TRUE.}

\item{useGinv}{logical. Whether to use generalized inverse for computing least-squares regression. 
Default FALSE, but may be helpful with very homogeneous datasets.}
}
\value{
sce object with velocities - see colData(sce) for the first two components (dX and dY),
and sce@metadata$vectors for all components.
}
\description{
Computes the inferred velocity of each cell in the dataset according to the topology supplied previously.
}
