% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_utilities.R
\name{simTopo}
\alias{simTopo}
\title{Simulate topology}
\usage{
simTopo(topo, ...)
}
\arguments{
\item{topo}{data.frame describing a network topology with columns Source, Target, and Type, where type 1
indicates activation and type 2 indicates inhibition.}

\item{...}{Other parameters for the sracipeSimulate function.}
}
\value{
sRACIPE object.
}
\description{
Simple wrapper function to simulate a topology using sracipeSimulate from the sRACIPE package. Returns
a RACIPE object.
}
