% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{smoothVector}
\alias{smoothVector}
\title{Compute a smoothed vector around a specific sample or gene expression state}
\usage{
smoothVector(sce, queryPoint, neighborhoodRadius, ...)
}
\arguments{
\item{sce}{sticcc object}

\item{queryPoint}{character or vector. Either a rowname of sce, or a vector to be compared to 
the normalized expression data.}

\item{neighborhoodRadius}{numeric. Proportion of the maximum pairwise distance within sce to find
neighbors to compute vectors for.}
}
\value{
A list containing the following elements: "start": the query point, "v1_x": v1 prediction
for the first PC, "v1_y": v1 prediction for the second PC, "v2_x", "v2_y", "net_x", "net_y", "rev_x",
"rev_y" denoting the other predictions in PC1 and PC2, "neighborhoodSize": number of nearby samples
used to compute the , "smoothingNumber":
}
\description{
Finds neighboring samples around the supplied point, computes vectors for each, and 
returns a distance-weighted average representing the smoothed vector for the region as if it 
were a grid point from computeGridVectors
}
