% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_utilities.R
\name{computeGrid}
\alias{computeGrid}
\title{Compute grid points for later smoothing}
\usage{
computeGrid(sce, grid.length = 30)
}
\arguments{
\item{sce}{viccc object. Must have already executed the function runPCA().}

\item{grid.length}{numeric. How many grid points should be in each side of the smoothed plot.
Default 30.}
}
\value{
viccc object with defined grid points in PCA space
}
\description{
Computes grid points for future steps that will smooth the inferred velocities
in PCA space.
}
