% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICCC_utilities.R
\name{computeGridVectors}
\alias{computeGridVectors}
\title{Grid-based smoothing of inferred velocities}
\usage{
computeGridVectors(sce, scalingFactor = NA)
}
\arguments{
\item{sce}{viccc object after computing inferred velocities.}

\item{scalingFactor}{numeric. Constant to multiply vectors by for visual clarity.}
}
\value{
viccc object with smoothed data in sce@metadata$grid.df
}
\description{
Uses inverse distance weighting to compute a smoothed grid of velocity vectors.
}
