% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{runPCA}
\alias{runPCA}
\title{Perform PCA on gene expression data}
\usage{
runPCA(sce, save = T, fname = NA)
}
\arguments{
\item{sce}{viccc object.
cells.}
}
\value{
modified viccc object containing pca transformed data in reducedDim(obj, "PCA") and pca parameters in obj@metadata$pca_data.
}
\description{
Returns a vector of cluster numbers corresponding to each cell in a dataset.
}
