% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STICCC_utilities.R
\name{runPCA}
\alias{runPCA}
\title{Perform PCA on gene expression data}
\usage{
runPCA(sce, save = T, overwrite = T, fname = NA)
}
\arguments{
\item{sce}{viccc object.}

\item{save}{logical. Whether to save pca results to file}

\item{overwrite}{logical. If true, pca file will overwrite preivous pca if one exists. Default TRUE. 
Set this to FALSE if you intend to use a specific PCA projection for further analysis}

\item{fname}{character. Where to save pca results. Default will be an Rds file named PCA_res in the data 
subdirectory of the topology directory.}
}
\value{
modified viccc object containing pca transformed data in reducedDim(obj, "PCA") and pca parameters in obj@metadata$pca_data.
}
\description{
Returns a vector of cluster numbers corresponding to each cell in a dataset.
}
